src/main/java/com/marinatedegg/sell/service/impl/ProductInfoServiceImpl.java
package com.marinatedegg.sell.service.impl;

import com.marinatedegg.sell.dao.ProductInfoDao;
import com.marinatedegg.sell.dataobject.ProductInfo;
import com.marinatedegg.sell.enums.ProductStatusEnum;
import com.marinatedegg.sell.service.ProductInfoService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ProductInfoServiceImpl implements ProductInfoService {

    @Autowired
    private ProductInfoDao dao;

    @Override
    public ProductInfo findOne(String productId) {
        return dao.findById(productId).get();
    }

    @Override
    public List<ProductInfo> findUpAll() {
        return dao.findByProductStatus(ProductStatusEnum.UP.getCode());
    }

    @Override
    public Page<ProductInfo> findAll(Pageable pageable) {
        return dao.findAll(pageable);
    }

    @Override
    public ProductInfo save(ProductInfo productInfo) {
        return dao.save(productInfo);
    }
}
